requirements.txt
external==0.0.1
keras==2.13.1
keras_nightly==2.5.0.dev2021032900
matplotlib==3.4.3
mxnet==2.0.0a0
numpy==1.22.3
onnx==1.11.0
onnx2keras==0.0.24
onnx2torch==1.5.4
onnx_tf==1.10.0
onnxmltools==1.11.1
onnxruntime==1.15.1
onnxruntime_gpu==1.12.1
opencv_python_headless==4.7.0.72
pandas==1.4.1
Pillow==9.5.0
protobuf==4.23.3
requests==2.24.0
scikit_learn==1.3.0
scipy==1.11.1
seaborn==0.11.2
tensorflow==2.8.0
tf2onnx==1.9.3
tflite2onnx==0.4.0
tflite==2.4.0
tflite-runtime==2.8.0
torch==1.11.0
torchvision==0.12.0
tvm==1.0.0
ddKS @ git+https://github.com/pnnl/DDKS@70f8822b95acf82b82b3ba768049765cd6b571c4
gpuparallel==0.2.2
